requirements.txt
langchain
python-dotenv
flask
flask_cors
elevenlabs
PyObjC
openai
pydantic==1.*
bs4
tiktoken
openai
pypdf
chromadb